Add backend discovery feature and simplify dock framework
