--- cmd/osdsdock/app/pool/pool.go
@@ -1,56 +1,0 @@
-// Copyright (c) 2016 Huawei Technologies Co., Ltd. All Rights Reserved.
-//
-//    Licensed under the Apache License, Version 2.0 (the "License"); you may
-//    not use this file except in compliance with the License. You may obtain
-//    a copy of the License at
-//
-//         http://www.apache.org/licenses/LICENSE-2.0
-//
-//    Unless required by applicable law or agreed to in writing, software
-//    distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
-//    WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
-//    License for the specific language governing permissions and limitations
-//    under the License.
-
-/*
-This module defines the initialization work about dock service: generate dock
-node information and record it to dock route service. Please DO NOT
-modify this file.
-
-*/
-
-package pool
-
-import (
-	"errors"
-	"log"
-
-	api "github.com/opensds/opensds/pkg/model"
-)
-
-func GetPool(poolId string) (*api.StoragePoolSpec, error) {
-	pools, err := readPoolsFromFile()
-	if err != nil {
-		log.Println("Could not read pool resource:", err)
-		return &api.StoragePoolSpec{}, err
-	}
-
-	for _, pool := range pools {
-		if poolId == pool.Id {
-			return &pool, nil
-		}
-	}
-
-	err = errors.New("Could not find any pool resource!")
-	return &api.StoragePoolSpec{}, err
-}
-
-func ListPools() (*[]api.StoragePoolSpec, error) {
-	pools, err := readPoolsFromFile()
-	if err != nil {
-		log.Println("Could not read pool resource:", err)
-		return &[]api.StoragePoolSpec{}, err
-	}
-
-	return &pools, nil
-}
