+++ pkg/dock/dock.go
@@ -47,45 +47,37 @@ func (d *DockHub) CreateVolume(name string, size int64) (*api.VolumeSpec, error)
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
 
+	log.Println("[Info] Calling volume plugin to create volume...")
+
 	//Call function of StoragePlugins configured by storage plugins.
-	result, err := d.VolumePlugin.CreateVolume(name, size)
-	if err != nil {
-		log.Println("Call plugin to create volume failed:", err)
-		return &api.VolumeSpec{}, err
-	} else {
-		return result, nil
-	}
+	return d.VolumePlugin.CreateVolume(name, size)
 }
 
 func (d *DockHub) GetVolume(volID string) (*api.VolumeSpec, error) {
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
 
+	log.Println("[Info] Calling volume plugin to get volume...")
+
 	//Call function of StoragePlugins configured by storage plugins.
-	result, err := d.VolumePlugin.GetVolume(volID)
-	if err != nil {
-		log.Println("Call plugin to get volume failed:", err)
-		return &api.VolumeSpec{}, err
-	} else {
-		return result, nil
-	}
+	return d.VolumePlugin.GetVolume(volID)
 }
 
 func (d *DockHub) DeleteVolume(volID string) error {
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
 
+	log.Println("[Info] Calling volume plugin to delete volume...")
+
 	//Call function of StoragePlugins configured by storage plugins.
-	if err := d.VolumePlugin.DeleteVolume(volID); err != nil {
-		log.Println("Call plugin to delete volume failed:", err)
-		return err
-	}
-	return nil
+	return d.VolumePlugin.DeleteVolume(volID)
 }
 
 func (d *DockHub) CreateVolumeAttachment(volID string, doLocalAttach, multiPath bool, hostInfo *api.HostInfo) (*api.VolumeAttachmentSpec, error) {
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
+
+	log.Println("[Info] Calling volume plugin to initialize connection...")
 
 	//Call function of StoragePlugins configured by storage plugins.
 	connInfo, err := d.VolumePlugin.InitializeConnection(volID, doLocalAttach, multiPath, hostInfo)
@@ -104,62 +96,58 @@ func (d *DockHub) UpdateVolumeAttachment(volID, host, mountpoint string) error {
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
 
+	log.Println("[Info] Calling volume plugin to attach volume...")
+
 	//Call function of StoragePlugins configured by storage plugins.
-	if err := d.VolumePlugin.AttachVolume(volID, host, mountpoint); err != nil {
-		log.Println("Call plugin to update volume attachment failed:", err)
-		return err
-	}
-	return nil
+	return d.VolumePlugin.AttachVolume(volID, host, mountpoint)
 }
 
 func (d *DockHub) DeleteVolumeAttachment(volID string) error {
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
 
+	log.Println("[Info] Calling volume plugin to detach volume...")
+
 	//Call function of StoragePlugins configured by storage plugins.
-	if err := d.VolumePlugin.DetachVolume(volID); err != nil {
-		log.Println("Call plugin to delete volume attachment failed:", err)
-		return err
-	}
-	return nil
+	return d.VolumePlugin.DetachVolume(volID)
 }
 
 func (d *DockHub) CreateSnapshot(name, volID, description string) (*api.VolumeSnapshotSpec, error) {
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
 
+	log.Println("[Info] Calling volume plugin to create snapshot...")
+
 	//Call function of StoragePlugins configured by storage plugins.
-	result, err := d.VolumePlugin.CreateSnapshot(name, volID, description)
-	if err != nil {
-		log.Println("Call plugin to create snapshot failed:", err)
-		return &api.VolumeSnapshotSpec{}, err
-	} else {
-		return result, nil
-	}
+	return d.VolumePlugin.CreateSnapshot(name, volID, description)
 }
 
 func (d *DockHub) GetSnapshot(snapID string) (*api.VolumeSnapshotSpec, error) {
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
 
+	log.Println("[Info] Calling volume plugin to get snapshot...")
+
 	//Call function of StoragePlugins configured by storage plugins.
-	result, err := d.VolumePlugin.GetSnapshot(snapID)
-	if err != nil {
-		log.Println("Call plugin to get snapshot failed:", err)
-		return &api.VolumeSnapshotSpec{}, err
-	} else {
-		return result, nil
-	}
+	return d.VolumePlugin.GetSnapshot(snapID)
 }
 
 func (d *DockHub) DeleteSnapshot(snapID string) error {
 	//Get the storage plugins and do some initializations.
 	d.VolumePlugin = plugins.InitVP(d.ResourceType)
 
+	log.Println("[Info] Calling volume plugin to delete snapshot...")
+
 	//Call function of StoragePlugins configured by storage plugins.
-	if err := d.VolumePlugin.DeleteSnapshot(snapID); err != nil {
-		log.Println("Call plugin to delete snapshot failed:", err)
-		return err
-	}
-	return nil
+	return d.VolumePlugin.DeleteSnapshot(snapID)
+}
+
+func (d *DockHub) ListPools() (*[]api.StoragePoolSpec, error) {
+	//Get the storage plugins and do some initializations.
+	d.VolumePlugin = plugins.InitVP(d.ResourceType)
+
+	log.Println("[Info] Calling volume plugin to list pools...")
+
+	//Call function of StoragePlugins configured by storage plugins.
+	return d.VolumePlugin.ListPools()
 }
